Products class

+++ DemoShopOnline/src/main/java/com/example/demo/repository/ProductsRepository.java
@@ -4,8 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Products;
 
-public interface ProductsRepository extends JpaRepository<Products, Long> {
-
-	
+public interface ProductsRepository extends JpaRepository<Products, Integer> {
 
 }
